+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- improved transpiler dumping GIL code on error.
+- GilToCsharp - a user output namespace is no longer required.
+  - https://github.com/StateSmith/StateSmith/issues/122
+
+----
+
 ## [0.8.1-alpha]
 ### Added
 - improve HistoryProcessor's use of expansionVarsPathProvider